Rename "specs" folder to "src" as required by the build script
