+++ packages/js/e2e-core-tests/src/activate-and-setup/activate.test.js
