+++ packages/js/e2e-core-tests/src/activate-and-setup/onboarding-tasklist.test.js
