+++ packages/js/e2e-core-tests/src/activate-and-setup/setup.test.js
