+++ packages/js/e2e-core-tests/src/api/external-product.test.js
