+++ packages/js/e2e-core-tests/src/api/telemetry.test.js
