+++ packages/js/e2e-core-tests/src/api/variable-product.test.js
