+++ packages/js/e2e-core-tests/src/merchant/wp-admin-analytics-page-loads.test.js
