+++ packages/js/e2e-core-tests/src/merchant/wp-admin-coupon-new.test.js
