+++ packages/js/e2e-core-tests/src/merchant/wp-admin-extensions-connect-wccom.test.js
