+++ packages/js/e2e-core-tests/src/merchant/wp-admin-order-apply-coupon.test.js
