+++ packages/js/e2e-core-tests/src/merchant/wp-admin-product-edit-details.test.js
