+++ packages/js/e2e-core-tests/src/merchant/wp-admin-product-new.test.js
