+++ packages/js/e2e-core-tests/src/merchant/wp-admin-settings-shipping-classes.test.js
