+++ packages/js/e2e-core-tests/src/merchant/wp-admin-settings-shipping-zones.test.js
