+++ packages/js/e2e-core-tests/src/shopper/front-end-checkout-coupons.test.js
