+++ packages/js/e2e-core-tests/src/shopper/front-end-order-email-receiving.test.js
