+++ packages/js/e2e-core-tests/src/shopper/front-end-product-browse-search-sort.test.js
