+++ packages/js/e2e-core-tests/src/shopper/front-end-single-product.test.js
