+++ packages/js/e2e-core-tests/src/shopper/front-end-variable-product-updates.test.js
