+++ packages/js/e2e-core-tests/src/utils/coupons.js
